+++ R/loo_compare.R
@@ -17,31 +17,34 @@
 #'
 #' @details
 #'   When comparing two fitted models, we can estimate the difference in their
-#'   expected predictive accuracy by the difference in `elpd_loo` or
-#'   `elpd_waic` (or multiplied by \eqn{-2}, if desired, to be on the
-#'   deviance scale).
+#'   expected predictive accuracy by the difference in
+#'   [`elpd_loo`][loo-glossary] or `elpd_waic` (or multiplied by \eqn{-2}, if
+#'   desired, to be on the deviance scale).
 #'
-#'   When using `loo_compare()`, the returned matrix will have one row per
-#'   model and several columns of estimates. The values in the `elpd_diff`
-#'   and `se_diff` columns of the returned matrix are computed by making
-#'   pairwise comparisons between each model and the model with the largest ELPD
-#'   (the model in the first row). For this reason the `elpd_diff` column
-#'   will always have the value `0` in the first row (i.e., the difference
-#'   between the preferred model and itself) and negative values in subsequent
-#'   rows for the remaining models.
+#'   When using `loo_compare()`, the returned matrix will have one row per model
+#'   and several columns of estimates. The values in the
+#'   [`elpd_diff`][loo-glossary] and [`se_diff`][loo-glossary] columns of the
+#'   returned matrix are computed by making pairwise comparisons between each
+#'   model and the model with the largest ELPD (the model in the first row). For
+#'   this reason the `elpd_diff` column will always have the value `0` in the
+#'   first row (i.e., the difference between the preferred model and itself) and
+#'   negative values in subsequent rows for the remaining models.
 #'
-#'   To compute the standard error of the difference in ELPD --- which should
-#'   not be expected to equal the difference of the standard errors --- we use a
-#'   paired estimate to take advantage of the fact that the same set of \eqn{N}
-#'   data points was used to fit both models. These calculations should be most
-#'   useful when \eqn{N} is large, because then non-normality of the
-#'   distribution is not such an issue when estimating the uncertainty in these
-#'   sums. These standard errors, for all their flaws, should give a better
-#'   sense of uncertainty than what is obtained using the current standard
-#'   approach of comparing differences of deviances to a Chi-squared
+#'   To compute the standard error of the difference in [ELPD][loo-glossary] ---
+#'   which should not be expected to equal the difference of the standard errors
+#'   --- we use a paired estimate to take advantage of the fact that the same
+#'   set of \eqn{N} data points was used to fit both models. These calculations
+#'   should be most useful when \eqn{N} is large, because then non-normality of
+#'   the distribution is not such an issue when estimating the uncertainty in
+#'   these sums. These standard errors, for all their flaws, should give a
+#'   better sense of uncertainty than what is obtained using the current
+#'   standard approach of comparing differences of deviances to a Chi-squared
 #'   distribution, a practice derived for Gaussian linear models or
 #'   asymptotically, and which only applies to nested models in any case.
 #'
+#' @seealso
+#' * The [FAQ page](https://mc-stan.org/loo/articles/online-only/faq.html) on
+#'   the __loo__ website for answers to frequently asked questions.
 #' @template loo-and-psis-references
 #'
 #' @examples
@@ -228,8 +231,7 @@ find_model_names <- function(x) {
       out_names[j] <- names4[j]
     }
   }
-
-  return(out_names)
+  out_names
 }