+++ tests/testthat/test_loo_subsampling.R
@@ -333,10 +333,7 @@ test_that("update.psis_loo_ss works as expected (compared with loo)", {
 
 context("loo_subsampling_approximations")
 
-test_that("elpd_loo_approximation works as expected", {
-
-
-  set.seed(123)
+geterate_test_elpd_dataset <- function() {
   N <- 10; K <- 10; S <- 1000; a0 <- 3; b0 <- 2
   p <- 0.7
   y <- rbinom(N, size = K, prob = p)
@@ -344,12 +341,22 @@ test_that("elpd_loo_approximation works as expected", {
   fake_posterior <- draws <- as.matrix(rbeta(S, a, b))
   fake_data <- data.frame(y,K)
   rm(N, K, S, a0, b0, p, y, a, b)
+
+  list(fake_posterior = fake_posterior, fake_data = fake_data)
+}
+
+test_elpd_loo_approximation <- function(cores) {
+  set.seed(123)
+  test_data <- geterate_test_elpd_dataset()
+  fake_posterior <- test_data$fake_posterior
+  fake_data <- test_data$fake_data
+
   llfun_test <- function(data_i, draws) {
     dbinom(data_i$y, size = data_i$K, prob = draws, log = TRUE)
   }
 
   # Compute plpd approximation
-  expect_silent(pi_vals <- loo:::elpd_loo_approximation(.llfun = llfun_test, data = fake_data, draws = fake_posterior, loo_approximation = "plpd", cores = 1))
+  expect_silent(pi_vals <- loo:::elpd_loo_approximation(.llfun = llfun_test, data = fake_data, draws = fake_posterior, loo_approximation = "plpd", cores = cores))
   # Compute it manually
   point <- mean(fake_posterior)
   llik <- dbinom(fake_data$y, size = fake_data$K, prob = point, log = TRUE)
@@ -358,7 +365,7 @@ test_that("elpd_loo_approximation works as expected", {
   expect_equal(abs(pi_vals)/sum(abs(pi_vals)), man_elpd_loo_approximation, tol = 0.00001)
 
   # Compute lpd approximation
-  expect_silent(pi_vals <- loo:::elpd_loo_approximation(.llfun = llfun_test, data = fake_data, draws = fake_posterior, loo_approximation = "lpd", cores = 1))
+  expect_silent(pi_vals <- loo:::elpd_loo_approximation(.llfun = llfun_test, data = fake_data, draws = fake_posterior, loo_approximation = "lpd", cores = cores))
   # Compute it manually
   llik <- numeric(10)
   for(i in seq_along(fake_data$y)){
@@ -369,7 +376,7 @@ test_that("elpd_loo_approximation works as expected", {
   expect_equal(abs(pi_vals)/sum(abs(pi_vals)), man_approx_loo_variable, tol = 0.00001)
 
   # Compute waic approximation
-  expect_silent(pi_vals_waic <- loo:::elpd_loo_approximation(.llfun = llfun_test, data = fake_data, draws = fake_posterior, loo_approximation = "waic", cores = 1))
+  expect_silent(pi_vals_waic <- loo:::elpd_loo_approximation(.llfun = llfun_test, data = fake_data, draws = fake_posterior, loo_approximation = "waic", cores = cores))
   expect_true(all(pi_vals > pi_vals_waic))
   expect_true(sum(pi_vals) - sum(pi_vals_waic) < 1)
 
@@ -379,11 +386,18 @@ test_that("elpd_loo_approximation works as expected", {
                                                             draws = fake_posterior,
                                                             loo_approximation = "tis",
                                                             loo_approximation_draws = 100,
-                                                            cores = 1))
+                                                            cores = cores))
   expect_true(all(pi_vals > pi_vals_tis))
   expect_true(sum(pi_vals) - sum(pi_vals_tis) < 1)
+}
+
+test_that("elpd_loo_approximation works as expected", {
+  test_elpd_loo_approximation(1)
 })
 
+test_that("elpd_loo_approximation with multiple cores", {
+  test_elpd_loo_approximation(2)
+})
 
 test_that("Test loo_approximation_draws", {
 
@@ -438,9 +452,9 @@ test_that("waic using delta method and gradient", {
     # fake_posterior <- mvtnorm::rmvnorm(n = S, mean = mu_n, sigma = solve(Lambda_n))
     colnames(fake_posterior) <- c("a", "b")
     fake_data <- data.frame(y, X)
-    save(fake_posterior, fake_data, file = test_path("normal_reg_waic_test_example.rda"))
+    save(fake_posterior, fake_data, file = test_path("data-for-tests/normal_reg_waic_test_example.rda"))
   } else {
-    load(file = test_path("normal_reg_waic_test_example.rda"))
+    load(file = test_path("data-for-tests/normal_reg_waic_test_example.rda"))
   }
 
   .llfun <- function(data_i, draws) {
@@ -492,9 +506,9 @@ test_that("waic using delta 2nd order method", {
     fake_posterior[,"sigma2"] <- sqrt(fake_posterior[,"sigma2"])
     colnames(fake_posterior) <- c("a", "b", "sigma")
     fake_data <- data.frame(y, X)
-    save(fake_posterior, fake_data, file = test_path("normal_reg_waic_test_example2.rda"), compression_level = 9)
+    save(fake_posterior, fake_data, file = test_path("data-for-tests/normal_reg_waic_test_example2.rda"), compression_level = 9)
   } else {
-    load(file = test_path("normal_reg_waic_test_example2.rda"))
+    load(file = test_path("data-for-tests/normal_reg_waic_test_example2.rda"))
   }
 
   .llfun <- function(data_i, draws) {
@@ -717,7 +731,7 @@ context("loo_subsampling cases")
 test_that("Test loo_subsampling and loo_approx with radon data", {
   skip_on_cran() # avoid going over time limit for tests
 
-  load(test_path("test_radon_laplace_loo.rda"))
+  load(test_path("data-for-tests/test_radon_laplace_loo.rda"))
   # Rename to spot variable leaking errors
   llfun_test <- llfun
   log_p_test <- log_p
@@ -890,10 +904,10 @@ test_that("Test the vignette", {
          stan_df, stan_df2,
          parameter_draws, parameter_draws_laplace, parameter_draws_2,
          log_p, log_g,
-         file = test_path("loo_subsample_vignette.rda"), compression_level = 9)
+         file = test_path("data-for-tests/loo_subsample_vignette.rda"), compression_level = 9)
 
   } else {
-    load(test_path("loo_subsample_vignette.rda"))
+    load(test_path("data-for-tests/loo_subsample_vignette.rda"))
   }
 
   set.seed(4711)